refactor(store): reorganize Redux store structure by moving slices to features directory
