+++ sticker-shop/src/App.jsx
@@ -3,7 +3,7 @@ import reactLogo from "./assets/react.svg";
 import viteLogo from "/vite.svg";
 import "./App.css";
 import { useDispatch, useSelector } from "react-redux";
-import { addItem } from "./slice/cartSlice";
+import { addItem } from "./store/features/cartSlice";
 
 function App() {
   const dispatch = useDispatch()
