+++ sticker-shop/src/store/features/productSlice.js
@@ -1,5 +1,5 @@
 import { createSlice, createEntityAdapter } from "@reduxjs/toolkit";
-import productApi from "../api/productApi"
+import productApi from "../../api/productApi"
 
 const productAdapter = createEntityAdapter()
